pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ─────────────────────────────────────────────────────────────────────────────
# SchoolOS – Python Project Definition
#
# Install everything:  pip install -e ".[dev]"
# Install prod only:   pip install -e .
# ─────────────────────────────────────────────────────────────────────────────

[project]
name        = "schoolos"
version     = "0.1.0"
description = "Multi-tenant AI Operating System for International Schools"
requires-python = ">=3.11"

dependencies = [
    # ── Web framework ───────────────────────────────────────────────────────
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",   # required for file uploads (CSV ingestion)

    # ── Database ────────────────────────────────────────────────────────────
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",           # async PostgreSQL driver
    "alembic>=1.13.0",           # database migrations

    # ── Settings & validation ───────────────────────────────────────────────
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",  # reads from .env automatically
    "python-dotenv>=1.0.0",

    # ── Cache & job queue ───────────────────────────────────────────────────
    "redis>=5.0.0",

    # ── Authentication ──────────────────────────────────────────────────────
    "python-jose[cryptography]>=3.3.0",  # JWT tokens
    "passlib[bcrypt]>=1.7.4",            # password hashing

    # ── Messaging ───────────────────────────────────────────────────────────
    "twilio>=9.0.0",    # WhatsApp + SMS
    "sendgrid>=6.11.0", # email

    # ── HTTP client (internal service calls) ────────────────────────────────
    "httpx>=0.27.0",

    # ── AI orchestration ────────────────────────────────────────────────────
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-groq>=0.1.0",     # Groq LLM provider (fast + free tier)

    # ── Timetable generation ─────────────────────────────────────────────
    "ortools>=9.10.0",           # Google constraint solver for scheduling
    "fpdf2>=2.7.0",              # PDF generation for printable timetables
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
# Tell hatch where the packages live
packages = ["shared", "services"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
